fix: resolve 5 CodeQL alerts in icctest V2 code
- cpp/enum-undefined-behavior: cast Severity comparisons to uint8_t
  in icctestParity.cpp and LinuxSandbox.cpp (2 alerts)
- cpp/enum-undefined-behavior: cast icColorSpaceSignature to
  icUInt32Number in IccQualityMetrics.h (1 alert)
- cpp/enum-undefined-behavior: cast Command comparison to uint8_t
  in ArgParser.cpp (1 alert)
- icc/narrow-loop-bound: widen nInput from uint8_t to int in
  ProfileView.cpp CLUT grid overflow check (1 alert)

V1: 813/813 tests pass, V2: all 3 targets build clean.

+++ iccanalyzer-lite/icctest/cli/src/LinuxSandbox.cpp
@@ -34,7 +34,7 @@ namespace {
 Severity maxFindingSeverity(const std::vector<Finding>& findings) {
     Severity worst = Severity::INFO;
     for (const auto& finding : findings) {
-        if (finding.level > worst) {
+        if (static_cast<uint8_t>(finding.level) > static_cast<uint8_t>(worst)) {
             worst = finding.level;
         }
     }
